@@ -16,18 +16,18 @@
   "dependencies": {
     "@3kles/kles-material-dialog": "^1.0.0",
     "@3kles/kles-material-dynamicforms": "^1.1.16",
-    "@angular/animations": "~12.0.1",
-    "@angular/cdk": "^12.0.1",
-    "@angular/common": "~12.0.1",
-    "@angular/compiler": "~12.0.1",
-    "@angular/core": "~12.0.1",
+    "@angular/animations": "~14.2.11",
+    "@angular/cdk": "^13.3.9",
+    "@angular/common": "~14.2.11",
+    "@angular/compiler": "~14.2.11",
+    "@angular/core": "~14.2.11",
     "@angular/flex-layout": "^9.0.0-beta.29",
-    "@angular/forms": "~12.0.1",
-    "@angular/material": "^12.0.1",
-    "@angular/material-moment-adapter": "^12.0.1",
-    "@angular/platform-browser": "~12.0.1",
-    "@angular/platform-browser-dynamic": "~12.0.1",
-    "@angular/router": "~12.0.1",
+    "@angular/forms": "~14.2.11",
+    "@angular/material": "^13.3.9",
+    "@angular/material-moment-adapter": "^13.3.9",
+    "@angular/platform-browser": "~14.2.11",
+    "@angular/platform-browser-dynamic": "~14.2.11",
+    "@angular/router": "~14.2.11",
     "@infor-up/m3-odin": "^4.0.0",
     "@ngx-translate/core": "^12.1.1",
     "@ngx-translate/http-loader": "^4.0.0",
@@ -46,8 +46,8 @@
     "@angular-devkit/build-angular": "^13.3.9",
     "@angular-devkit/build-ng-packagr": "^0.901.15",
     "@angular/cli": "~13.3.9",
-    "@angular/compiler-cli": "~12.0.1",
-    "@angular/language-service": "~12.0.1",
+    "@angular/compiler-cli": "~14.2.11",
+    "@angular/language-service": "~14.2.11",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/lodash": "^4.14.168",
@@ -64,6 +64,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.2.4"
+    "typescript": "~4.8.4"
   }
 }
