@@ -1,14 +1,14 @@
 import { KlesFieldAbstract } from '@3kles/kles-material-dynamicforms';
-import { FormGroup } from '@angular/forms';
+import { UntypedFormGroup } from '@angular/forms';
 import { KlesTreeColumnConfig } from '../../../models/columnconfig.model';
 
 export abstract class AbstractCell {
     column: KlesTreeColumnConfig;
-    row: FormGroup;
+    row: UntypedFormGroup;
     field: KlesFieldAbstract;
-    group: FormGroup;
+    group: UntypedFormGroup;
 
-    formatIndentation(node: FormGroup, step: number = 5): string {
+    formatIndentation(node: UntypedFormGroup, step: number = 5): string {
         return '&nbsp;'.repeat(node.value._status.depth * step);
     }
 
