@@ -1,7 +1,7 @@
-import { FormGroup } from '@angular/forms';
+import { UntypedFormGroup } from '@angular/forms';
 import { Observable } from 'rxjs';
 
 export interface ISelection {
-    select?: (selected: boolean, group: FormGroup) => Observable<{ selected: boolean; indeterminate: boolean }>;
+    select?: (selected: boolean, group: UntypedFormGroup) => Observable<{ selected: boolean; indeterminate: boolean }>;
     selectAll?: (selected: boolean, filters: { [key: string]: any }) => Observable<{ selected: boolean; indeterminate: boolean }>;
 }
