@@ -1,25 +1,25 @@
 import { IKlesFieldConfig } from '@3kles/kles-material-dynamicforms';
-import { FormArray, FormGroup } from '@angular/forms';
+import { UntypedFormArray, UntypedFormGroup } from '@angular/forms';
 import { Observable } from 'rxjs';
 import { KlesColumnConfig } from './columnconfig.model';
 
 export interface IKlesCellFieldConfig extends IKlesFieldConfig {
-    executeAfterChange?: (property?: string, row?: any, group?: FormGroup | FormArray) => Observable<any>;
+    executeAfterChange?: (property?: string, row?: any, group?: UntypedFormGroup | UntypedFormArray) => Observable<any>;
 }
 export interface IChangeHeaderFooterCell {
     column: KlesColumnConfig;
-    group: FormGroup | FormArray;
+    group: UntypedFormGroup | UntypedFormArray;
 }
 
 export interface IChangeCell {
     column: KlesColumnConfig;
     row: any;
-    group: FormGroup | FormArray;
+    group: UntypedFormGroup | UntypedFormArray;
     response?: any;
 }
 
 export interface IChangeLine {
-    group: FormGroup | FormArray;
+    group: UntypedFormGroup | UntypedFormArray;
     row: any;
     value: any;
 }
