@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { AbstractControl, FormGroup } from '@angular/forms';
+import { AbstractControl, UntypedFormGroup } from '@angular/forms';
 import { PageEvent } from '@angular/material/paginator';
 import { SafeStyle } from '@angular/platform-browser';
 import { IChangeCell, IChangeHeaderFooterCell, IChangeLine } from '../models/cell.model';
@@ -38,7 +38,7 @@ export abstract class AbstractKlesTableService {
     abstract onPageChange(e: PageEvent);
 
     //Manage Record
-    abstract addRecord(record, index?: number): FormGroup;
+    abstract addRecord(record, index?: number): UntypedFormGroup;
     abstract deleteRecord(record);
     abstract updateRecord(record, options?: { emitEvent: boolean, onlySelf: boolean });
 
