@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core";
-import { AbstractControl, FormGroup } from "@angular/forms";
+import { AbstractControl, UntypedFormGroup } from "@angular/forms";
 import { AbstractKlesTableService } from "../abstracttable.service";
 
 @Injectable({
@@ -8,7 +8,7 @@ import { AbstractKlesTableService } from "../abstracttable.service";
 export abstract class AbstractKlesTreeTableService extends AbstractKlesTableService {
     protected table: any;
     abstract getDepthDataAccessor(item: AbstractControl, property: string): number;
-    abstract getParentDataAccessor(item: FormGroup, property: string): AbstractControl;
+    abstract getParentDataAccessor(item: UntypedFormGroup, property: string): AbstractControl;
     abstract onLineOpen(e: any);
     abstract onLineClose(e: any);
     abstract updateRow(record: any, options?: { emitEvent: boolean; onlySelf: boolean; });
